Replace locations page
