+++ src/components/Pages.tsx
@@ -6,11 +6,10 @@ import { Sidebar } from "./UI/molecules/Sidebar";
 import { Gases } from "./UI/organisms/Gases";
 import { Home } from "./UI/organisms/Home";
 import { Incidents } from "./UI/organisms/Incidents";
-import { Locations } from "./UI/organisms/Locations";
 import { Login } from "./UI/organisms/Login";
 import { UserAccount } from "./UI/organisms/UserAccount";
 import { API_URL, getCurrentUser } from "../index";
-import { LocationsNext } from "./UI/organisms/LocationsNext";
+import { Locations } from "./UI/organisms/Locations";
 
 const useStyles = makeStyles({
   content: {
@@ -65,12 +64,6 @@ export const Pages: React.FC = () => {
                   path="/locations"
                   title="Blackline Safety | Locations"
                   component={Locations}
-                />
-                <Page
-                  exact
-                  path="/locationsNext"
-                  title="Blackline Safety | Locations"
-                  component={LocationsNext}
                 />
                 <Page
                   exact
